Creating marketing route group
Created a marketing folder that organizes the pages related to marketing. Created layout.tsx to render the layout for all pages in the marketing route group. Imported a button from shadcn library. Added the font to the public folder and imported the font locally. Appended tailwind with imported font with cn import. Created a reusable logo component. Created a private folder named _components to opt all components out of routing. Added the navbar and footer from the private component folder and imported them into the index.